requirements.txt
vapi-python
requests
python-dotenv
newsapi-python
yfinance
aiohttp 